# Implement a faster sort algorithm