@nagisa 2016-12-06
Since it is quite a major rewrite of the sorting routine, and I see that we have no test guaranteeing stability, I would very much like a test that ensures this invariant.

Something along the lines of

```
let x = [0, 0, 0, 1, 1, 1, 2, 2, 2, 3, 3, 3];
let y = // array/vector/slice of references to each element of x. References compare the underlying value
y.shuffle();
let z = y.clone();
y.sort();
// assert that the addresses in y and z are ordered in the same order for each block of values.
```