@stjepang 2016-12-06
Actually, we do have a sort guaranteeing stability. It's called `test_sort_stability`. However, it tests sort on short arrays only, essentially verifying correctness of just insertion sort. I'll bump the lengths up a bit.