@stjepang 2016-12-07
I addressed all requests for changes and improved comments in the code.