@mark-i-m 2016-12-07
Re: https://github.com/rust-lang/rust/pull/38192#discussion_r91120627, sorting a collection of ZSTs

For what its worth, I think the _only_ correct answer is to make it a nop. ZSTs are types with exactly one value. Thus, given a collection `x` of values of type `T`, for some ZST `T`, and two arbitrary values in the collection `x[i]` and `x[j]`, we know that `x[i] == x[j]`, so the only correct behavior (for a stable sort) is to preserve the exact order of the input collection, right?