@mark-i-m 2016-12-07
Ah, I see.

Thanks

On Dec 7, 2016 4:10 AM, "Stjepan Glavina" <notifications@github.com> wrote:

> *@stjepang* commented on this pull request.
> ------------------------------
>
> In src/libcollections/slice.rs
> <https://github.com/rust-lang/rust/pull/38192>:
>
> > @@ -1087,7 +1086,7 @@ impl<T> [T] {
>      /// elements.
>      ///
>      /// This sort is stable and `O(n log n)` worst-case but allocates
> -    /// approximately `2 * n`, where `n` is the length of `self`.
> +    /// approximately `n / 2`, where `n` is the length of `self`.
>
> Saying "allocates O(n / 2)" is equivalent to "allocates less than n / 2 *
> K, for some constant K".
>
> Saying "allocates approximately n / 2" is equivalent to "allocates
> between n / 2 - K and n / 2 + K", for some constant K".
>
> The latter is more precise and therefore a better description of memory
> usage.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rust/pull/38192>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AIazwFQoAPKJio1ylmgBO8x7imia05Txks5rFoYYgaJpZM4LFWL1>
> .
>