@gnzlbg 2016-12-07
Is there a case for which this implementation performs worse than the old one?