@stjepang 2016-12-07
@gnzlbg I don't think so. Let's discuss a few cases:

1. Small arrays (below the insertion sort threshold). I tried sorting big & small structures on ascending & descending & random inputs. The sorts are always either tied, or the new sort wins.

2. Medium sized arrays (just above the insertion sort threshold). One might think that additional allocation caused by `runs` is troubling, but turns out it isn't. I tried poking at various cases. Again, the new sort wins in all of them, except a few in which they are tied.

3. Large arrays. The new sort destroys it on partially sorted arrays, so such cases are no contest. The old sort is strongest on random inputs, and even then it's consistently behind the new sort. There is a test called `sort_large_random_expensive`, which sorts a random array using a very costly comparison function. I've examined the total number of comparisons required using the old vs new algorithm, and it turns out the new one performs a slightly smaller number of comparisons. It's no wonder it wins with a slight edge in that test, too.

Sorting depends on structure size, type of input, comparison function, your platform, etc. I can't guarantee the new sort **always** wins, but haven't found a case where it wouldn't. If it ever loses, I'm sure that would be a race lost by a very small margin.