@gnzlbg 2016-12-07
@stjepang In practice, most data to be sorted has some structure and is rarely truly random. I think it would be worth it to check the benchmarking patterns of boost::sort, libc++ std::sort, [cpp-sort](https://github.com/Morwenn/cpp-sort), ... (the licenses are permissive). For example, libc++-std::sort is benchmarked using the following patterns:

![p8k6kxe](https://cloud.githubusercontent.com/assets/904614/20973576/678efb52-bc99-11e6-8eda-e428044d38df.png)

Besides these patterns, the following patterns are also common:

- [ascending/descending sawtooth](https://github.com/Morwenn/cpp-sort/blob/master/benchmarks/distributions.h#L213)
- [alternating/alternating N](https://github.com/Morwenn/cpp-sort/blob/master/benchmarks/distributions.h#L247)

See also [this Boost.Sort issue](https://github.com/boostorg/sort/issues/11#issuecomment-23044336). 

IMO the new implementation doesn't need to be better than the old one in all of these cases, but I still would like to be sure that it doesn't blow up in any (e.g. libc++'s std::sort turned out to be N^2 for some particular patterns, see [llvm's bugzilla](https://llvm.org/bugs/show_bug.cgi?id=20837) for the pattern and the fix).