@bluss 2016-12-07
ok, can you explain what the two performance listings mean and which numbers are for which implementation? Important information is missing like: larger is better(?)