@Stebalien 2016-12-07
@Morwenn rust could do this through specialization and a unstable-sort-safe marker trait.